[#313] Legacy API 호출 코드 삭제
