--- Qapple/Qapple/SourceCode/Data/Repository/DTO/Answer/DeleteAnswerDTO.swift
@@ -1,12 +1,0 @@
-//
-//  DeleteAnswerDTO.swift
-//  Qapple
-//
-//  Created by Simmons on 1/22/25.
-//
-
-import Foundation
-
-struct DeleteAnswerDTO: Codable {
-    let answerId: Int
-}
