--- Qapple/Qapple/SourceCode/Data/Repository/DTO/BulletinBoard/DeleteBoardDTO.swift
@@ -1,12 +1,0 @@
-//
-//  DeleteBoardDTO.swift
-//  Qapple
-//
-//  Created by Simmons on 1/22/25.
-//
-
-import Foundation
-
-struct DeleteBoardDTO: Codable {
-    let boardId: Int
-}
