--- Qapple/Qapple/SourceCode/Data/Repository/DTO/Members/MyPageDTO.swift
@@ -1,14 +1,0 @@
-//
-//  MyPageDTO.swift
-//  Qapple
-//
-//  Created by Simmons on 1/22/25.
-//
-
-import Foundation
-
-struct MyPageDTO: Codable {
-    let nickname: String
-    let profileImage: String?
-    let joinDate: String
-}
