--- Qapple/Qapple/SourceCode/Data/Repository/DTO/Members/ResignDTO.swift
@@ -1,12 +1,0 @@
-//
-//  ResignDTO.swift
-//  Qapple
-//
-//  Created by 김민준 on 2/5/25.
-//
-
-import Foundation
-
-struct ResignDTO: Codable {
-    let memberId: Int
-}
