--- Qapple/Qapple/SourceCode/Data/Repository/DTO/Reports/BoardCommentReportsDTO.swift
@@ -1,17 +1,0 @@
-//
-//  BoardCommentReportsDTO.swift
-//  Qapple
-//
-//  Created by Simmons on 1/22/25.
-//
-
-import Foundation
-
-struct BoardCommentReportsDTO: Codable {
-    let boardCommentReportId: Int
-}
-
-struct BoardCommentReportsRequest: Codable {
-    let boardCommentId: Int
-    let boardCommentReportType: String
-}
