--- Qapple/Qapple/SourceCode/Data/Repository/DTO/Reports/BoardReportsDTO.swift
@@ -1,17 +1,0 @@
-//
-//  BoardReportsDTO.swift
-//  Qapple
-//
-//  Created by Simmons on 1/22/25.
-//
-
-import Foundation
-
-struct BoardReportsDTO: Codable {
-    let boardReportId: Int
-}
-
-struct BoardReportsRequest: Codable {
-    let boardId: Int
-    let boardReportType: String
-}
