--- Qapple/Qapple/SourceCode/Data/Repository/DependencyKey/BoardCommentKey.swift
@@ -1,49 +1,0 @@
-//
-//  BoardCommentKey.swift
-//  Qapple
-//
-//  Created by Simmons on 1/22/25.
-//
-
-import Foundation
-
-extension QappleRepository {
-    
-    /// 게시글 댓글 리스트 조회
-    static func makeFetchBoardCommentList() -> (_ boardId: Int, _ threshold: Int?) async throws -> ([BoardComment], QappleAPI.PaginationInfo) {
-        return { boardId, threshold in
-            let url = try QappleAPI.BoardComment.list(boardId: boardId, threshold: threshold, pageSize: 25).url()
-            let response: BaseResponse<BoardCommentsDTO> = try await NetworkService.shared.get(url: url)
-            return response.result.toEntityWithThreshold
-        }
-    }
-    
-    /// 게시글 댓글 삭제
-    static func makeDeleteBoardComment() -> (_ commentId: Int) async throws -> DeleteBoardCommentsDTO {
-        return { commentId in
-            let url = try QappleAPI.BoardComment.delete(commentId: commentId).url()
-            let response: BaseResponse<DeleteBoardCommentsDTO> = try await NetworkService.shared.delete(url: url)
-            return response.result
-        }
-    }
-    
-    /// 게시글 댓글 생성
-    static func makePostBoardComment() -> (_ boardId: Int, _ comment: String) async throws -> PostBoardCommentsDTO {
-        return { boardId, comment in
-            let url = try QappleAPI.BoardComment.post(boardId: boardId).url()
-            let requestBody: PostBoardCommentsRequest = PostBoardCommentsRequest(comment: comment)
-            let response: BaseResponse<PostBoardCommentsDTO> = try await NetworkService.shared.post(url: url, body: requestBody)
-            return response.result
-        }
-    }
-    
-    /// 게시글 댓글 좋아요/취소 토글
-    static func makeLikeBoardComment() -> (_ commentId: Int) async throws -> LikeBoardCommentsDTO {
-        return { commentId in
-            let url = try QappleAPI.BoardComment.like(commentId: commentId).url()
-            let requestBody: LikeBoardCommentsRequest = LikeBoardCommentsRequest(commentId: commentId)
-            let response: BaseResponse<LikeBoardCommentsDTO> = try await NetworkService.shared.patch(url: url, body: requestBody)
-            return response.result
-        }
-    }
-}
