--- Qapple/Qapple/SourceCode/Data/Repository/DependencyKey/NotificationKey.swift
@@ -1,20 +1,0 @@
-//
-//  NotificationKey.swift
-//  Qapple
-//
-//  Created by Simmons on 1/22/25.
-//
-
-import Foundation
-
-extension QappleRepository {
-    
-    /// 알림 리스트 조회
-    static func makeFetchNotificationList() -> (_ threshold: Int?) async throws -> ([QappleNotification], QappleAPI.PaginationInfo) {
-        return { threshold in
-            let url = try QappleAPI.Notification.list(threshold: threshold, pageSize: 25).url()
-            let response: BaseResponse<NotificationsDTO> = try await NetworkService.shared.get(url: url)
-            return response.result.toEntityWithThreshold
-        }
-    }
-}
