--- Qapple/Qapple/SourceCode/Data/Repository/DependencyKey/QuestionKey.swift
@@ -1,29 +1,0 @@
-//
-//  QuestionKey.swift
-//  Qapple
-//
-//  Created by Simmons on 1/21/25.
-//
-
-import Foundation
-
-extension QappleRepository {
-    
-    /// 모든 질문 조회
-    static func makeFetchQuestionList() -> (_ threshold: String?) async throws -> ([Question], QappleAPI.TotalCount, QappleAPI.PaginationInfo) {
-        return { threshold in
-            let url = try QappleAPI.Question.list(threshold: threshold, pageSize: 25).url()
-            let response: BaseResponse<QuestionsDTO> = try await NetworkService.shared.get(url: url)
-            return response.result.toEntityWithInfo
-        }
-    }
-    
-    /// 메인 질문 조회
-    static func makeFetchMainQuestionList() -> () async throws -> Question {
-        return {
-            let url = try QappleAPI.Question.listOfMain.url()
-            let response: BaseResponse<MainQuestionDTO> = try await NetworkService.shared.get(url: url)
-            return response.result.toEntity
-        }
-    }
-}
