--- Qapple/Qapple/SourceCode/Data/Repository/DependencyKey/ReportsKey.swift
@@ -1,41 +1,0 @@
-//
-//  ReportsKey.swift
-//  Qapple
-//
-//  Created by Simmons on 1/22/25.
-//
-
-import Foundation
-
-extension QappleRepository {
-    
-    /// 게시판 신고
-    static func makeReportsBoard() -> (_ boardId: Int, _ boardReportType: String) async throws -> BoardReportsDTO {
-        return { boardId, boardReportType in
-            let url = try QappleAPI.Reports.board.url()
-            let requestBody: BoardReportsRequest = BoardReportsRequest(boardId: boardId, boardReportType: boardReportType)
-            let response: BaseResponse<BoardReportsDTO> = try await NetworkService.shared.post(url: url, body: requestBody)
-            return response.result
-        }
-    }
-    
-    /// 게시판 댓글 신고
-    static func makeReportsBoardComment() -> (_ boardCommentId: Int, _ boardCommentReportType: String) async throws -> BoardCommentReportsDTO {
-        return { boardCommentId, boardCommentReportType in
-            let url = try QappleAPI.Reports.boardComment.url()
-            let requestBody: BoardCommentReportsRequest = BoardCommentReportsRequest(boardCommentId: boardCommentId , boardCommentReportType: boardCommentReportType)
-            let response: BaseResponse<BoardCommentReportsDTO> = try await NetworkService.shared.post(url: url, body: requestBody)
-            return response.result
-        }
-    }
-    
-    /// 답변 신고
-    static func makeReportsAnswer() -> (_ answerId: Int, _ reportType: String) async throws -> AnswerReportsDTO {
-        return { answerId, reportType in
-            let url = try QappleAPI.Reports.answer.url()
-            let requestBody: AnswerReportsRequest = AnswerReportsRequest(answerId: answerId, reportType: reportType)
-            let response: BaseResponse<AnswerReportsDTO> = try await NetworkService.shared.post(url: url, body: requestBody)
-            return response.result
-        }
-    }
-}
