--- Qapple/Qapple/SourceCode/Data/Repository/DependencyKey/TokenKey.swift
@@ -1,20 +1,0 @@
-//
-//  TokenKey.swift
-//  Qapple
-//
-//  Created by Simmons on 1/22/25.
-//
-
-import Foundation
-
-extension QappleRepository {
-    
-    /// 토큰 재발급
-    static func makeRefreshToken() -> () async throws -> RefreshToken {
-        return {
-            let url = try QappleAPI.Token.refresh.url()
-            let response: BaseResponse<RefreshTokenDTO> = try await NetworkService.shared.get(url: url)
-            return response.result.toEntity
-        }
-    }
-}
