--- Qapple/Qapple/SourceCode/Entity/QuestionOfMain.swift
@@ -1,14 +1,0 @@
-//
-//  MainQuestion.swift
-//  Qapple
-//
-//  Created by Simmons on 1/21/25.
-//
-
-import Foundation
-
-struct QuestionOfMain: Identifiable, Equatable {
-    let id: Int
-    let title: String
-    var isAnswered: Bool
-}
